Conflicts:
	webapp/src/edu/cornell/mannlib/vitro/webapp/filters/VitroRequestPrep.java
	webapp/src/edu/cornell/mannlib/vitro/webapp/filters/WebappDaoFactorySDBPrep.java
	webapp/src/edu/cornell/mannlib/vitro/webapp/filters/WebappDaoFactorySparqlPrep.java
	webapp/web/WEB-INF/ontologies/app/menu.n3
	webapp/web/WEB-INF/ontologies/app/menuload/displayTBOX.n3

+++ webapp/src/edu/cornell/mannlib/vitro/webapp/controller/freemarker/DeletePropertyController.java
@@ -9,6 +9,8 @@ import javax.servlet.http.HttpServletResponse;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 
+import edu.cornell.mannlib.vitro.webapp.auth.permissions.SimplePermission;
+import edu.cornell.mannlib.vitro.webapp.auth.requestedAction.Actions;
 import edu.cornell.mannlib.vitro.webapp.beans.DataProperty;
 import edu.cornell.mannlib.vitro.webapp.beans.DataPropertyStatement;
 import edu.cornell.mannlib.vitro.webapp.beans.Individual;
@@ -17,9 +19,11 @@ import edu.cornell.mannlib.vitro.webapp.controller.VitroRequest;
 import edu.cornell.mannlib.vitro.webapp.controller.freemarker.responsevalues.RedirectResponseValues;
 import edu.cornell.mannlib.vitro.webapp.controller.freemarker.responsevalues.ResponseValues;
 import edu.cornell.mannlib.vitro.webapp.controller.freemarker.responsevalues.TemplateResponseValues;
+import edu.cornell.mannlib.vitro.webapp.dao.ModelAccess;
+import edu.cornell.mannlib.vitro.webapp.dao.ModelAccess.FactoryID;
 import edu.cornell.mannlib.vitro.webapp.dao.WebappDaoFactory;
 import edu.cornell.mannlib.vitro.webapp.edit.n3editing.VTwo.EditConfigurationUtils;
-import edu.cornell.mannlib.vitro.webapp.edit.n3editing.processEdit.EditN3Utils;
+import edu.cornell.mannlib.vitro.webapp.edit.n3editing.VTwo.N3EditUtils;
 import edu.cornell.mannlib.vitro.webapp.web.URLEncoder;
 /*
  * Custom deletion controller to which deletion requests from default property form are sent. May be replaced 
@@ -27,17 +31,11 @@ import edu.cornell.mannlib.vitro.webapp.web.URLEncoder;
  */
 public class DeletePropertyController extends FreemarkerHttpServlet {
     private static final Log log = LogFactory.getLog(DeletePropertyController.class);
-  
-    //since forwarding from edit Request dispatch for now
-   //TODO: Check what required actions would make sense here
-    //public final static Actions REQUIRED_ACTIONS = new Actions(new ManageMenus());
-    
-   
-    /*
-     *  @Override
-    protected Actions requiredActions(VitroRequest vreq) {
-    	return REQUIRED_ACTIONS;
-    }*/
+ 
+    @Override
+	protected Actions requiredActions(VitroRequest vreq) {
+    	return SimplePermission.DO_FRONT_END_EDITING.ACTIONS ;
+	}
 
     @Override
     protected ResponseValues processRequest(VitroRequest vreq) {
@@ -106,7 +104,7 @@ public class DeletePropertyController extends FreemarkerHttpServlet {
 		//This is the standard mechanism but note that datapropStmtDelete uses wdf with user aware
 
 		//DataProperty prop = EditConfigurationUtils.getDataProperty(vreq);
-		String editorUri = EditN3Utils.getEditorUri(vreq);
+		String editorUri = N3EditUtils.getEditorUri(vreq);
 		WebappDaoFactory wdf = vreq.getWebappDaoFactory().getUserAwareDaoFactory(editorUri);
 		DataProperty prop = wdf.getDataPropertyDao().getDataPropertyByURI(
 				EditConfigurationUtils.getPredicateUri(vreq));
@@ -169,38 +167,38 @@ public class DeletePropertyController extends FreemarkerHttpServlet {
 		
 	}
 
-
-	
-
-
 	//process object property
     private void processObjectProperty(VitroRequest vreq) {
     	ObjectProperty prop = EditConfigurationUtils.getObjectProperty(vreq);
-    	
+    	    	
     	//if this property is true, it means the object needs to be deleted along with statement
-    	if(prop.getStubObjectRelation())
+    	//while the second test is to see if a different object uri (i.e. not the direct objet of the predicate)
+    	//needs to be deleted
+    	if(prop.getStubObjectRelation() || hasDeleteObjectUri(vreq))
     	{
     		deleteObjectIndividual(vreq);
     	}
     	
-    	deleteObjectPropertyStatement(vreq);
+    	if(!hasDeleteObjectUri(vreq)) {
+    		deleteObjectPropertyStatement(vreq);
+    	}
 		
     }
     
     private void deleteObjectPropertyStatement(VitroRequest vreq) {
 		WebappDaoFactory wdf = vreq.getWebappDaoFactory();
-    	String objectUri = EditConfigurationUtils.getObjectUri(vreq);
+		String objectUri = EditConfigurationUtils.getObjectUri(vreq);
 		String subjectUri = EditConfigurationUtils.getSubjectUri(vreq);
 		String predicateUri = EditConfigurationUtils.getPredicateUri(vreq);
 		//delete object property statement
 		wdf.getPropertyInstanceDao().deleteObjectPropertyStatement(subjectUri, predicateUri, objectUri);
 	}
 
-	private Individual getObjectIndividualForStubRelation(VitroRequest vreq, String objectUri) {
+	private Individual getObjectIndividualForDeletion(VitroRequest vreq, String objectUri) {
     
     	Individual object = EditConfigurationUtils.getIndividual(vreq, objectUri);
     	if(object == null) {
-    		WebappDaoFactory wadf = (WebappDaoFactory) vreq.getSession().getServletContext().getAttribute("webappDaoFactory");
+			WebappDaoFactory wadf = ModelAccess.on(vreq.getSession().getServletContext()).getWebappDaoFactory();
     		object = wadf.getIndividualDao().getIndividualByURI(objectUri);
     	}
     	
@@ -210,15 +208,29 @@ public class DeletePropertyController extends FreemarkerHttpServlet {
     
     private void deleteObjectIndividual(VitroRequest vreq) {
     	String objectUri = EditConfigurationUtils.getObjectUri(vreq);
-    	Individual object = getObjectIndividualForStubRelation(vreq, objectUri);
+    	if(hasDeleteObjectUri(vreq)) {
+    		//if a different individual needs to be deleted, get that uri instead
+    		objectUri = getDeleteObjectUri(vreq);
+    	}
+    	Individual object = getObjectIndividualForDeletion(vreq, objectUri);
     	if(object != null) {
-    		log.warn("Deleting individual " + object.getName() + "since property has been set to force range object deletion");
+    		log.warn("Deleting individual " + object.getName() + "since property has been set to force range object deletion or has been set to delete a specific object");
     		WebappDaoFactory wdf = vreq.getWebappDaoFactory();
     		wdf.getIndividualDao().deleteIndividual(object);
     	} else {
     		//TODO: Throw error?
     		log.error("could not find object as request attribute or in model " + objectUri);
     	}
+    }
+    
+    //This checks if the object uri is not the individual to be deleted but another individual connected
+    private String getDeleteObjectUri(VitroRequest vreq) {
+    	return (String) vreq.getParameter("deleteObjectUri");
+    }
+    
+    private boolean hasDeleteObjectUri(VitroRequest vreq) {
+    	String deleteObjectUri = getDeleteObjectUri(vreq);
+    	return (deleteObjectUri != null && !deleteObjectUri.isEmpty());
     }
